refactor: split node renderer height measurements
